loss/acc fixed for tpu

+++ transformer_bert.py
@@ -309,7 +309,7 @@ def loss_function(real, pred):
     mask = tf.math.logical_not(tf.math.equal(real, 0))
     loss_ = loss_object(real, pred)
 
-    mask = tf.cast(mask, dtype=loss_.dtype)
+    mask = tf.cast(mask, dtype=tf.float64)
     loss_ *= mask
     loss_reduced = tf.reduce_sum(loss_)/tf.reduce_sum(mask)
     return loss_reduced
